% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_functions.R
\name{extractFirstSnpPositionByGene}
\alias{extractFirstSnpPositionByGene}
\title{Extract First Snp Position By Gene}
\usage{
extractFirstSnpPositionByGene(snpAnnot, return_matrix = TRUE)
}
\description{
Extract First Snp Position By Gene
}
