% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{filterSamples}
\alias{filterSamples}
\title{Filter Samples}
\usage{
filterSamples(samples, ratios, heteroScores, annotations, cmpSamples = NULL,
  considerThisSNP = NULL)
}
\description{
Filter Samples
}
