% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessXCI.R
\name{guessXCI}
\alias{guessXCI}
\title{Guess X Chromosome Inactivation}
\usage{
guessXCI(file, annotCol = 7, depth = 10, countsToCall = 5, total = TRUE,
  callheteroThr = 0.2, plot = T)
}
\description{
Guess X Chromosome Inactivation
}
