% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{readGATK.AllelicImbalance}
\alias{readGATK.AllelicImbalance}
\alias{readFromVCF}
\title{Read allelic expression file}
\usage{
readGATK.AllelicImbalance(file, annotCol = 7)

readFromVCF(file.vcf)
}
\arguments{
\item{file}{a file name}

\item{annotCol}{the column number where the annotations ends}

\item{file.vcf}{a vcf file name}
}
\value{
a list with
  \item{alleleCounts}{each row is an allele (SNP) expression value}
  \item{alleleAnnot}{allele annotations}
  \item{positionUID}{uinque id for snp allele}
}
\description{
Read allelic expression file

Read allelic expression from VCF
}
