% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_guessedLOI}
\alias{plot_guessedLOI}
\title{Make a plot of the LOI genes}
\usage{
plot_guessedLOI(heteroGene, geneOrder = NULL, minToSee = 1, flatTo = 5,
  gaps_row = NULL, gaps_col = NULL, addThisMono = NULL, removeME = NULL,
  plot = T)
}
\description{
Make a plot of the LOI genes
}
