% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plotAllelicRatios}
\alias{plotAllelicRatios}
\title{Make a plot of the allelic ratios}
\usage{
plotAllelicRatios(obj, samples = NULL, na.rm = F,
  removeSpecificIds = NULL)
}
\description{
Make a plot of the allelic ratios
}
