% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{processSexualData}
\alias{processSexualData}
\alias{processNonSexualData}
\title{Create X Y linked SNP's allelic expression}
\usage{
processSexualData(gatk_ai, countsToCall, total, callheteroThr,
  species = c("hsapien", "mmusculus"), removeNonDBsnp = TRUE)

processNonSexualData(gatk_ai, countsToCall, total, callheteroThr,
  species = c("hsapien", "mmusculus"), removeNonDBsnp = TRUE)
}
\arguments{
\item{gatk_ai}{allelic imbalance from \code{readGATK.AllelicImbalance}}

\item{countsToCall}{how many counts to the alternative allele to call}

\item{total}{use tota as denominator in the allele ratios}

\item{callheteroThr}{the threshold to call hetorozygous}

\item{species}{either hsapiens or mmusculus. Choose the species to remove pseudoAutosomal regions}

\item{removeNonDBsnp}{should I remove the non DBsnp SNPs}
}
\value{
a list with
  \item{genes}{genes related object for visualization}
  \item{snps}{snps related object for visualization}
  \item{geneSummary}{tabular summary of the genes}
  \item{snpSummary}{tabular summary of the SNPs}
}
\description{
Create X Y linked SNP's allelic expression

Create Non Sexual linked SNP's allelic expression
}
