% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_functions.R
\name{ratioInString}
\alias{ratioInString}
\title{Create a string representing the ration in string}
\usage{
ratioInString(alleles, idx, total = FALSE)
}
\description{
Create a string representing the ration in string
}
