% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{removeHg38PsudosomalRegion}
\alias{removeHg38PsudosomalRegion}
\title{Remove Human pseudosomal regions}
\usage{
removeHg38PsudosomalRegion(obj, withChr = "FALSE")
}
\description{
Remove Human pseudosomal regions
}
