% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_functions.R
\name{removeM38PsudosomalRegion}
\alias{removeM38PsudosomalRegion}
\title{Remove Mouse Pseudosomal Regions}
\usage{
removeM38PsudosomalRegion(obj, withChr = "FALSE")
}
\description{
Remove Mouse Pseudosomal Regions
}
